You should be able to send data to plugin, to update translations
Added implementation of saving received translation data from SNS.
Update Language checkbox to be an array without keys in React script.

Add AWS package to the plugin 23mb... (just a quick copy paste of their
downloaded version of their package, which means the WHOLE plugin!).

Added a plugin settings page, so we can store various things we need to
send to the BTS to have LW work in different ways.