Changes to have the javascript work properly
We can now send data to wp admin ajax, and parse the returning response
properly, so the language states are updated.
Changes for how we send data to backend.
Changes to how the metabox looks.
Added a small "sent" status, after the data is sent to backend, and a
response is returned.

+ other small updates as well.