Handle missing groups and do not create empty trans-units
No longer adding empty trans-units, if the given ACF field has array
content instead of string data. This is typically seen on the images.
Perhaps we should think about adding this "later" though?

When ACF data was being parsed, the content was skipped and it seems
the group attributes was still parsed. I've remove these to see if it
helps.