Handle line breaks
Line breaks cannot be sent as \r etc, as LW seems to be unable
to use these correctly.
LW has supplied a list of chars we need to use instead.
Added const's with the line break arrays (LW and normal).
Added two methods to convert to and from LW linebreaks.

BTS-65