Update post slug, when a post is translated
A post slug (post_name) is now set, when the article is translated.
NOTE: The slug is ALWAYS overwritten, if the title is changed.

Also setting the post_author, when copying the post.
This is somewhat related to BTS-70, but it's not the proper one,
as it's the author from the taxonomy ACF fields.

BTS-70, BTS-68