Add override author to translated posts
We need to copy the custom author to the translated
posts. The author is inside a custom metabox.

Added a new method to set post meta on a post. Note that
the "old" post meta with the same key should be removed
before the new one can be set.

Altered how we set the translation state, now using the
new method as well.

Bumped version to 0.2.17

BTS-62