Getting started
The new Willow Translation Service Plugin is in progress.
I have added a metabox similar to what we had in the FZ
plugin, but there is no javascript and data is directly
saved when we save the post.
Added some styling and an empty javascript file.

You can now save data to the post, but language copies
are not created.